+++ sdk/storage/azure-storage-queue/azure/storage/queue/aio/_queue_client_async.py
@@ -375,6 +375,7 @@ class QueueClient(AsyncStorageAccountHostsMixin, QueueClientBase):
             warnings.warn(
                 "TypeError when calling message_encode_policy.configure. \
                 It is likely missing the encryption_version parameter. \
+                Consider updating your encryption information/implementation. \
                 Retrying without encryption_version."
             )
             self._config.message_encode_policy.configure(
@@ -614,8 +615,10 @@ class QueueClient(AsyncStorageAccountHostsMixin, QueueClientBase):
                 )
             except TypeError:
                 warnings.warn(
-                    "message_encode_policy.configure does not accept encryption_version parameter.",
-                    DeprecationWarning
+                    "TypeError when calling message_encode_policy.configure. \
+                    It is likely missing the encryption_version parameter. \
+                    Consider updating your encryption information/implementation. \
+                    Retrying without encryption_version."
                 )
                 self._config.message_encode_policy.configure(
                     self.require_encryption,