+++ sdk/storage/azure-storage-queue/CHANGELOG.md
@@ -3,9 +3,11 @@
 ## 12.4.0b1 (Unreleased)
 
 ### Features Added
-- Added a new version of client-side encryption for Queue messages (version 2.0) which utilizes AES-GCM-256 encryption.
-If you are currently using client-side encryption, it is **highly recommended** to switch to version 2.0. The encryption
-version can be specified on any client constructor via the `encryption_version` keyword (`encryption_version='2.0'`).
+- Introduced version 2.0 of client-side encryption for Queue messages which utilizes AES-GCM-256 encryption.
+Version 1.0 is now deprecated and no longer considered secure. If you are using client-side encryption, it is
+**highly recommended** that you update to version 2.0.
+The encryption version can be specified on any client constructor via the `encryption_version`
+keyword (i.e. `encryption_version='2.0'`).
 
 ## 12.3.0 (2022-05-09)